fix: move `addDotSeparator` into lib/utils since route.ts can only export route functions
