+++ app/api/[area]/[code]/boundary/route.ts
@@ -2,27 +2,10 @@ import { NextRequest } from 'next/server'
 import { z } from 'zod'
 import { get } from 'https'
 import { Params, paramSchema } from './schema'
+import { addDotSeparator } from '@/lib/utils'
 
 const ghRawBaseUrl =
   'https://raw.githubusercontent.com/fityannugroho/idn-area-boundary/main/data'
-
-/**
- * Add dot separator to the code.
- * - 4 digits (e.g. 9603) becomes 5 digits (e.g. 96.03)
- * - 6 digits (e.g. 960301) becomes 8 digits (e.g. 96.03.01)
- * - 10 digits (e.g. 9603011001) becomes 12 digits (e.g. 96.03.01.1001)
- */
-export function addDotSeparator(code: string) {
-  const codeLength = code.length
-
-  if (codeLength === 4) return `${code.slice(0, 2)}.${code.slice(2)}`
-  if (codeLength === 6)
-    return `${code.slice(0, 2)}.${code.slice(2, 4)}.${code.slice(4)}`
-  if (codeLength === 10)
-    return `${code.slice(0, 2)}.${code.slice(2, 4)}.${code.slice(4, 6)}.${code.slice(6)}`
-
-  return code
-}
 
 export async function GET(
   request: NextRequest,
