+++ lib/utils.ts
@@ -1,6 +1,24 @@
 import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 
+/**
+ * Add dot separator for the area code.
+ * - 4 digits (e.g. 9603) becomes 5 digits (e.g. 96.03)
+ * - 6 digits (e.g. 960301) becomes 8 digits (e.g. 96.03.01)
+ * - 10 digits (e.g. 9603011001) becomes 12 digits (e.g. 96.03.01.1001)
+ */
+export function addDotSeparator(code: string) {
+  const codeLength = code.length
+
+  if (codeLength === 4) return `${code.slice(0, 2)}.${code.slice(2)}`
+  if (codeLength === 6)
+    return `${code.slice(0, 2)}.${code.slice(2, 4)}.${code.slice(4)}`
+  if (codeLength === 10)
+    return `${code.slice(0, 2)}.${code.slice(2, 4)}.${code.slice(4, 6)}.${code.slice(6)}`
+
+  return code
+}
+
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
